fix warning

+++ src/Open3D/Container/Tensor.h
@@ -62,7 +62,7 @@ public:
            const Device& device = Device("CPU:0"))
         : Tensor(shape, dtype, device) {
         // Check number of elements
-        if (init_vals.size() != shape_.NumElements()) {
+        if (static_cast<int64_t>(init_vals.size()) != shape_.NumElements()) {
             utility::LogError(
                     "Tensor initialization values' size {} does not match the "
                     "shape {}",
